requirements_tests.txt
pytest
coverage
torchvision
